requirements.txt
langchain>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.20
fastapi>=0.104.0
uvicorn>=0.23.2
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.4.2
httpx>=0.25.0